Extracted BEP package
